Updating Documentation
https://trello.com/c/gmcaVGUy/38-updating-documentation

We lost a few marks in our snapshot due to lack of documentation, and in class I remember it being mentioned that our README should be thorough enough that it would bring someone new to the project up to speed, so I'm planning on sprucing it up a bit

basically providing more user-friendly instructions on how to set up a dev. environment, maybe pasting that UI architecture plan, providing links to our programmings practices, architecture, etc.